Refactor: Changes to schema dtos, entities, guards and resolvers based on requirements

+++ src/product-rent-details/product-rent.entity.ts
@@ -1,4 +1,4 @@
-import { ObjectType, Field, Int, Float } from '@nestjs/graphql';
+import { ObjectType, Field, Int } from '@nestjs/graphql';
 import { ProductEntity } from 'src/products/product.entity';
 
 @ObjectType()
@@ -9,7 +9,7 @@ export class ProductRentEntity {
   @Field(() => Int)
   productId: number;
 
-  @Field(() => Float)
+  @Field(() => Int)
   price: number;
 
   @Field(() => Date)
